services/analytics/requirements.txt
# Analytics Service Dependencies

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
tortoise-orm[asyncpg]==0.20.0
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Scheduling
apscheduler==3.10.4

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Mathematical Operations
scipy==1.11.4

# Caching
aiocache==0.12.2

# HTTP Client for notifications
aiohttp==3.9.1

# Email support
aiosmtplib==3.0.1

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1